requirements.txt
pdfminer.six
nltk
requests
textract